requirements.txt
openai 
python-dotenv
spotipy